requirements.txt
flask
flask-cors
spotipy
python-dotenv
requests
openai
cryptography
google-generativeai